requirements.txt
# Core ML libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0  # Para cargar modelos guardados

# Data loading
ucimlrepo>=0.0.7  # Para descargar datasets de UCI

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web application
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0  # Para validación de datos

# Database
pymongo>=4.4.0
motor>=3.0.0  # Cliente async de MongoDB para FastAPI

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # Para TestClient de FastAPI